Decompose common file functionality for reuse
This is stored in ui-kit so that CMS can reuse it as well
